LPS-77992 Rename portal-osgi-web-jasper-plugins to portal-osgi-web-jasper-plugins-taglib.
