+++ modules/apps/static/portal-osgi-web/portal-osgi-web-jasper-plugins-taglib/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Portal OSGi Web Jasper Plugins Taglib
+Bundle-SymbolicName: com.liferay.portal.osgi.web.jasper.plugins.taglib
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.osgi.web.jasper.plugins.taglib
+Fragment-Host: com.liferay.portal.osgi.web.servlet.jsp.compiler
+Import-Package: !org.apache.jasper.compiler.tagplugin
