+++ modules/apps/static/portal-osgi-web/portal-osgi-web-jasper-plugins-taglib/src/main/resources/WEB-INF/tagPlugins.xml
@@ -3,18 +3,18 @@
 <tag-plugins>
 	<tag-plugin>
 		<tag-class>com.liferay.taglib.core.ChooseTag</tag-class>
-		<plugin-class>com.liferay.portal.osgi.web.jasper.plugins.ChooseTagPlugin</plugin-class>
+		<plugin-class>com.liferay.portal.osgi.web.jasper.plugins.taglib.ChooseTagPlugin</plugin-class>
 	</tag-plugin>
 	<tag-plugin>
 		<tag-class>com.liferay.taglib.core.IfTag</tag-class>
-		<plugin-class>com.liferay.portal.osgi.web.jasper.plugins.IfTagPlugin</plugin-class>
+		<plugin-class>com.liferay.portal.osgi.web.jasper.plugins.taglib.IfTagPlugin</plugin-class>
 	</tag-plugin>
 	<tag-plugin>
 		<tag-class>com.liferay.taglib.core.OtherwiseTag</tag-class>
-		<plugin-class>com.liferay.portal.osgi.web.jasper.plugins.OtherwiseTagPlugin</plugin-class>
+		<plugin-class>com.liferay.portal.osgi.web.jasper.plugins.taglib.OtherwiseTagPlugin</plugin-class>
 	</tag-plugin>
 	<tag-plugin>
 		<tag-class>com.liferay.taglib.core.WhenTag</tag-class>
-		<plugin-class>com.liferay.portal.osgi.web.jasper.plugins.WhenTagPlugin</plugin-class>
+		<plugin-class>com.liferay.portal.osgi.web.jasper.plugins.taglib.WhenTagPlugin</plugin-class>
 	</tag-plugin>
 </tag-plugins>
