--- modules/apps/static/portal-osgi-web/portal-osgi-web-jasper-plugins/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay Portal OSGi Web Jasper Plugins
-Bundle-SymbolicName: com.liferay.portal.osgi.web.jasper.plugins
-Bundle-Version: 1.0.4
-Export-Package: com.liferay.portal.osgi.web.jasper.plugins
-Fragment-Host: com.liferay.portal.osgi.web.servlet.jsp.compiler
-Import-Package: !org.apache.jasper.compiler.tagplugin
