+++ portal-impl/src/portal.properties
@@ -6248,7 +6248,7 @@
     # "/WEB-INF/lib" directory to be statically included in all bundles.
     #
     module.framework.static.jars=\
-        com.liferay.portal.osgi.web.jasper.plugins.jar,\
+        com.liferay.portal.osgi.web.jasper.plugins.taglib.jar,\
         util-taglib.jar
 
     #
